refactoring for viewmodels
