+++ ContactBook/app/build.gradle
@@ -11,7 +11,7 @@ android {
 
     defaultConfig {
         applicationId "com.example.contactbook"
-        minSdk 26
+        minSdk 23
         targetSdk 31
         versionCode 1
         versionName "1.0"
