+++ ContactBook/app/src/main/java/com/example/contactbook/ui/viewModels/ContactViewModel.kt
@@ -1,15 +1,14 @@
 package com.example.contactbook.ui.viewModels
 
-import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.contactbook.data.entities.Contact
 import com.example.contactbook.data.repositories.implementations.ContactRepository
 import com.example.contactbook.data.services.InputValidationService
-import com.example.contactbook.data.services.abstractions.IInputValidationService
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 @HiltViewModel
@@ -31,17 +30,23 @@ class ContactViewModel @Inject constructor(
         }
     }
 
-    fun getContacts(ownerId: String): LiveData<List<Contact>> {
-        return contactRepository.getContacts(ownerId)
+    suspend fun getContact(id: String): Contact = withContext(Dispatchers.IO) {
+        return@withContext contactRepository.getContact(id)
     }
 
     fun checkInputValidation(name: String, instagram: String, phone: String): Array<Boolean> {
         return inputValidationService.addContactInputValidation(name, instagram, phone)
     }
-    
-    fun deleteContacts(ownerId: String) {
+
+    fun editContact(contact: Contact) {
         viewModelScope.launch(Dispatchers.IO) {
-            contactRepository.deleteContacts(ownerId)
+            contactRepository.editContact(contact)
+        }
+    }
+
+    fun deleteContact(id: String) {
+        viewModelScope.launch(Dispatchers.IO) {
+            contactRepository.deleteContact(id)
         }
     }
 }
