+++ ContactBook/app/src/main/res/layout/fragment_contact_detail.xml
@@ -85,20 +85,4 @@
                 app:layout_constraintTop_toBottomOf="@+id/instagram_TV" />
         </androidx.appcompat.widget.LinearLayoutCompat>
     </androidx.cardview.widget.CardView>
-
-
-    <ImageButton
-        android:id="@+id/edit_Btn"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:backgroundTint="@color/white"
-        android:minHeight="48dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.03"
-        app:srcCompat="@drawable/ic_edit" />
-
 </androidx.constraintlayout.widget.ConstraintLayout>
