+++ ContactBook/app/src/main/res/layout/fragment_login.xml
@@ -35,7 +35,7 @@
                 android:text="@string/login_login_text"
                 android:textColor="@android:color/white"
 
-                android:textSize="15sp"
+                android:textSize="15dp"
                 app:layout_constraintBottom_toBottomOf="parent"
 
                 app:layout_constraintEnd_toEndOf="parent"
@@ -60,7 +60,7 @@
 
                 android:backgroundTint="@android:color/white"
                 android:textColor="@color/main"
-                android:textSize="15sp"
+                android:textSize="15dp"
                 app:layout_constraintBottom_toTopOf="@+id/emailError"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toEndOf="@+id/login_fragment_button"
@@ -162,7 +162,7 @@
                 android:text="@string/login_fragment"
 
                 android:textColor="@android:color/white"
-                android:textSize="15sp"
+                android:textSize="15dp"
                 app:layout_constraintBottom_toTopOf="@+id/emailError"
                 app:layout_constraintEnd_toStartOf="@+id/register_fragment_button"
                 app:layout_constraintStart_toStartOf="parent"
