+++ ContactBook/app/src/main/res/layout/fragment_register.xml
@@ -38,7 +38,7 @@
                 android:backgroundTint="@android:color/white"
                 android:text="@string/login_fragment"
                 android:textColor="@color/main"
-                android:textSize="15sp"
+                android:textSize="15dp"
 
                 app:layout_constraintBottom_toTopOf="@+id/nameError"
                 app:layout_constraintEnd_toStartOf="@+id/register_fragment_button"
@@ -65,7 +65,7 @@
                 android:text="@string/register_fragment"
                 android:textColor="@android:color/white"
 
-                android:textSize="15sp"
+                android:textSize="15dp"
                 app:layout_constraintBottom_toTopOf="@+id/nameError"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toEndOf="@+id/login_fragment_button"
@@ -89,7 +89,7 @@
                 android:text="@string/register_register_confirm_button_text"
 
                 android:textColor="@android:color/white"
-                android:textSize="15sp"
+                android:textSize="15dp"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintHorizontal_bias="0.512"
@@ -102,17 +102,12 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="40dp"
-
                 android:layout_marginEnd="40dp"
                 android:layout_marginBottom="20dp"
                 android:theme="@style/MyTextInputLayout_overlay"
                 app:boxStrokeWidth="2dp"
                 app:counterEnabled="true"
-
-
                 app:counterMaxLength="20"
-                app:counterTextColor="@color/grey"
-
                 app:errorEnabled="true"
                 app:helperText="@string/register_name_helper_text"
                 app:helperTextTextColor="@color/grey"
